Big update to how results are calculated.
Big update to how results are calculated. Each metric is linearly scaled across the likely range from prior seasons. These are then scaled by weights, combined into a total weighted factor per team, and then applied to the calculated score. This is an attempt to simulate the relative strengths of each team and how that differential might affect their score. This is a major change to the prediction algorithm.